updates:  1. buttons- “add to cart” and “buy now”. 2. out of stock 3. price to be double 4. message for email success in 2 pages- contact and home
finish!

+++ Online-Ceramics-Store/Models/Product.cs
@@ -12,13 +12,13 @@ namespace Online_Ceramics_Store.Models
         public string description { get; set; }
         public string information { get; set; }
 		public int quantity_purchased { get; set; }
-        public int UpdatedPrice { get; set; }
+        public double UpdatedPrice { get; set; }
 
         public int stock_quantity { get; set; }
         public int category_id { get; set; }
         public int insale { get; set; }
         public int percent { get; set; }
-        public int price { get; set; }
+        public double price { get; set; }
 
 
 	}
